requirements.txt
altgraph==0.17
APScheduler==3.6.3
astroid==2.6.2
autopep8==1.5.7
backports.entry-points-selectable==1.1.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.9.3
cachetools==4.2.2
certifi==2021.5.30
charset-normalizer==2.0.2
colorama==0.4.4
configparser==5.0.2
crayons==0.4.0
distlib==0.3.2
filelock==3.0.12
future==0.18.2
idna==3.2
importlib-metadata==4.6.4
isort==5.9.1
lazy-object-proxy==1.6.0
lxml==4.6.3
mccabe==0.6.1
numpy==1.21.0
pefile==2021.5.24
Pillow==8.3.1
pipenv==2021.5.29
platformdirs==2.2.0
pycodestyle==2.7.0
pyfiglet==0.8.post1
pyinstaller==4.5.1
pyinstaller-hooks-contrib==2021.2
pylint==2.9.3
python-telegram-bot==13.7
pytube==11.0.0
pytz==2021.1
pywin32-ctypes==0.2.0
requests==2.26.0
selenium==3.141.0
six==1.16.0
soupsieve==2.2.1
termcolor==1.1.0
toml==0.10.2
tornado==6.1
typed-ast==1.4.3
typing-extensions==3.10.0.0
tzdata==2021.1
tzlocal==3.0
urllib3==1.26.6
virtualenv==20.7.2
virtualenv-clone==0.5.6
webdriver-manager==3.4.2
wrapt==1.12.1
zipp==3.5.0
